Back/requirements.txt
asgiref==3.8.1
boto3==1.38.32
botocore==1.38.32
click==8.2.1
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
django-storages==1.14.4
gunicorn==23.0.0
h11==0.16.0
imbalanced-learn==0.12.4
imblearn==0.0
jmespath==1.0.1
joblib==1.4.2
mysqlclient==2.2.7
numpy==2.3.0
packaging==25.0
pandas==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.13.0
scikit-learn==1.4.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.3
whitenoise>=6.5.0